+++ crates/warpui_core/Cargo.toml
@@ -8,10 +8,6 @@ license = "MIT"
 
 [features]
 traces = []
-# Purely additive: enables the TUI backend (views, elements, presenter,
-# runtime) alongside the always-compiled GUI machinery. The default (GUI)
-# build is unaffected, and enabling this feature anywhere in a dependency
-# graph does not change GUI consumers (feature unification is harmless).
 tui = ["dep:crossterm", "dep:unicode-segmentation", "dep:unicode-width"]
 # Feature enabled only when app is compiled for integration tests.
 integration_tests = ["dep:minimp4", "dep:openh264"]
