+++ crates/warpui_core/src/core/view/mod.rs
@@ -164,7 +164,3 @@ pub trait TypedActionView {
         ActionAccessibilityContent::default()
     }
 }
-
-#[cfg(test)]
-#[path = "mod_test.rs"]
-mod tests;
